Install MkDocs
This will replace the `multicloudpaas.github.gds` site which currently uses
Jekyll. During alpha we've found that it hasn't been so well suited to a
documentation site (compared to a website/blog). We need a hierarchy of
content and a very simple process for adding/removing content.

MkDocs gives us those things. It's very similar to Sphinx which is used for
the GOV.UK opsmanual, but simpler still. The one thing that we lose by not
using Jekyll is the automatic build process in GitHub Pages. I intend to
replace this with Read The Docs, which has native support for MkDocs,
because I don't anticipate any of our content needing to be private.