Working practices: port from old opsmanual
Take information about commit messages, pull requests, dashboard, and
Pivotal integration from:

- https://github.gds/multicloudpaas/multicloudpaas.github.gds/blob/master/development.md

I've reworded these to emphasise *why* we're doing it and grouped together
everything that is related to pull requests.

With apologies to @annshipman that wrote most of this content the first
time. I tried to import the history but it wasn't very easy/clear.