Add a diagram illustrating our development process
Add a diagram we can use to explain the process for getting changes into
production. The audience for this are people within the PaaS team and other
people within GDS, such as those involved in the security accreditation
process.

Currently the diagram must be generated from source and the results be
committed.