Add development process to working practices
Add the development process diagram and a bullet point explanation of our
process to the description of team working practices for the benefit of PaaS
team members and other people within GDS, such as accreditors, who need to
understand how we safeguard production.

This is a description of how the process works now; it does not include planned
improvements to the process, which should be added only when they are
implemented.