Reduce entities in diagram for legibility
Previously the deployers and environments were shown as separate entities for
CI, staging and production. The number of entities and the fact that mkdocs has
quite a narrow fixed width meant that this was quite hard to read.

The deployer and environments were folded into one and triggering / deployment
were made into one activity.