Only upgrade one thing at a time
We recently attempted to upgrade BOSH, CF and stemcells in one PR:
https://www.pivotaltracker.com/story/show/118272895

The result of this was that BOSH failed to deploy and left things in an
inconsistent state, but not every time. We saw this once while working on the
story, but decided it was probably caused by problems with that development
environment, which had shown other problems.

When the pull request was reviewd it became clear that this problem occurred
commonly after the upgrade, but not always. After quite a lot of debugging we
thought, but weren't certain that it was caused by the BOSH upgrade, so we
prepared a branch that upgraded CF and the stemcells, not BOSH and raised an
issue upstream describing the problems we saw with the BOSH upgrade.

When we tested this branch we saw a new error where BOSH would timeout talking
to agents. Because we'd seen this error in other environments before the
upgrade we thought, but weren't certain, that this was nothing to do with
either the CF or stemcell upgrade.

In order to be sure that this was the case we then prepared yet another branch
with an upgrade to CF only and tested this. This was successful.

All this took several days, occupied the time of three people and a review
slot. If we'd put each upgrade in its own pull request this would have not
occupied so much review time.

Splitting them into 3 separate changes would mean:
- If the PR runs perfectly and is accepted with no changes, 3x the testing time
  in review
- If we discovered a problem with one of the changes it could be shelved easily
- The story would not be occupy so much time in review when things did not go
  smoothly