What testing is required
During the course of a recent story to upgrade concourse we realised that there
are 3 different start points for the upgrade that we could test and the most
important one is starting from current master, as it's closest to what will
happen in prod.

However, our ability to use our development environments to make and test
changes is also important, so we should test what happens after our nightly
autodelete-cloudfoundry pipeline has run.