Add ADR for AWS credentials
To cover work across the following stories:

IAM roles and profiles:
- https://www.pivotaltracker.com/story/show/109806354
- https://www.pivotaltracker.com/story/show/110246622
- https://www.pivotaltracker.com/story/show/111138770
- https://www.pivotaltracker.com/story/show/110238314
- https://www.pivotaltracker.com/story/show/110311736
- https://www.pivotaltracker.com/story/show/111303414
- https://www.pivotaltracker.com/story/show/111778374
- https://www.pivotaltracker.com/story/show/111523754

IP restrictions:
- https://www.pivotaltracker.com/story/show/109625018

This topic has come up in a few conversations recently because we've been
experiencing problems with the EC2 instance metadata service. So it would be
useful to state clearly why we don't want to use pre-generated keys.

I picked ADR03 because that was the next available when I first started
writing this months ago and it's still available today.