Add reminder to check Diego job ordering
In story #115884129[1] the Diego components were moved off the
colocated VM and onto separate VMs. The VMs have been placed in
a specific order to match that used by the Diego development
team in their testing. This adds a reminder to check we always
have the right order after upgrades.

[1] https://www.pivotaltracker.com/story/show/115884129