Link to spreadsheet as canonical accounts list
This spreadsheet is now the canonical reference for what accounts a team
member *could* have and the URLs for accessing each one. By putting it in a
Google Doc we can be more comfortable about linking to semi-sensitive
information like our AWS account aliases. It will also become our reference
for starters and leavers, which should be more consistent than the Pivotal
Tracker tasks that we have at the moment.