Remove uses of eg, ie and etc as per the styleguide
See https://www.gov.uk/guidance/style-guide/a-to-z-of-gov-uk-style
In the absence of automation for this sort of thing we should try to
be consistent to avoid style differences across our documentation
corpus. Unfortunately we can't avoid being arbitrary in our selection
of rules, but we can do better than we were doing.