Add IPSec debugging guide
We're about to implement the IPSec encryption, between some of our
nodes. It may so happen, that one will be in need of debugging the IPSec
connectivity. Wrong time in the wrong place, I suppose.

In order to make their task simpler, we'd like to document some of the
techniques we've used while implementing the IPSec.