You should ask people to prioritise follow-up work
We spread the duty of accepting stories across the team, so that not just the
product manager, tech leads and tech arch could do it.

This was successful in the sense of having fewer stories waiting in acceptance,
but caused a new problem in that the people who planned work for a sprint no
longer had a good idea of follow-up work needed to be done after existing work.

As a result, we agreed at our team retro on 2016-12-13 that we should make it
clear that part of the job of approving work should be to make one of these
people aware of any follow-up work that should be prioritised.