How to deal with uncontactable org managers
In https://gaap.deskpro.com/agent/#app.tickets,t:3627 we discovered that
someone who was the sole org manager for a trial organisation had left that
organisation.

We had previously amended our user docs to encourage having at least two org
managers, to minimise the risk of this happening.

We agreed who would be able to decide who to put in their place and that it
would have to be done on a case by case basis, because we don't know of a
repeatable process to discover who the replacement should be.