Test rotating credentials when doing CF upgrades
New releases can introduce new passwords or different uses of existing
passwords. It's possible that these may not be easy to rotate without
causing downtime. We'd like to know if this is the case before we next need
to use the credential rotation in anger (e.g. leaked secrets).

This was originally part of a follow-up story but I've pulled it out after a
discussion in knowledge share because we're likely to do one or more CF
upgrades before we play the follow-up to rotate more credentials.