Moved BOSH connection scripts to be in the guide
https://github.com/alphagov/paas-cf/pull/1023 added several scripts for
connecting to BOSH. We then added a guide to the team manual that told
us how to use them https://github.com/alphagov/paas-team-manual/pull/137

After more discussion we've decided to move these scripts to be in the
guide. There are no automated tests of the scripts, they may well break,
and thus it is best to have them inline as more of a getting-started
guide.

@leeporte identified in approval that there were issues with targeting
BOSH via a tunnel, and that creating a new VM as a gateway using the
state bucket's id_rsa did not work in his environment. This also
addresses these issues.