Added guide to connecting to Concourse and BOSH
We want instructions ready for connecting to BOSH if we can't use our
normal Concourse task. Until now such instructions have lived in the
paas-cf README.md, but this has grown to a large size and the team
manual is a more suitable location.

This commit adds those instructions, covering a range of severities of
incident:

 * How to SSH or bosh-cli if Concourse tasks are unusable

 * How to SSH or bosh-cli if the Concourse VM is unusable

 * How to SSH or bosh-cli if we cannot create new VMs