Backup instructions for using `bosh ssh`
In approval @alext found that `bosh ssh` did not work because its SSH
connections were happening from outside of the VPC and were blocked by
our security rules.

The ideal fix would be to change our `ssh` commands to setup a SOCKS
proxy and use `BOSH_ALL_PROXY` as described by
https://bosh.io/docs/cli-tunnel.html. However this is only available on
bosh_cli V2 and we are on V1.

With that convenient approach ruled out, in this commit we add
instructions for manually specifying a gateway in a `bosh ssh` command.