Refactor restoring CF database docs
These rename `db_hostname` to `db_endpoint` because it is more consistent
with AWS terminology for this things.

We also provide instructions for dev environments because an operator may
want to practice in dev first. In dev we do not retain snapshots, and we
may not have the time to wait for snapshots to be generated by the normal
backup process.