Encourage diverse pairing practices
Pairing means a lot of different things to different people. Our
Working Practices didn't make clear that in each pair you have a wide
choice of how to work and that it's more important that the way you
pair works for both of you than to pair in any particular way.