Remove domain in links during link checking
We were having issues with HTMLProofer when checking cannonical
links[1], because the links simply did not exist yet.

But HTMLProofer allows pass an option `:url_swap` to apply a
ruby gsub to any link. We use it to remove the canonical server
name when checking the links.

The hostname is consumed from the config object itself.

[1] https://github.com/gjtorikian/html-proofer/issues/170