Inject google site search using js
Inject the Google Site search just before the navigation title.

This code inject on load the HTML element <gcse:search>[1] and
loads the google JS to configure the search. The custom
search engine has been configured using:
https://cse.google.co.uk/cse/ with my GDS google account (hector).

For any change, we can simply generate a new search in a different
account.

The search results will open as a 'popup' over the page.

[1] https://developers.google.com/custom-search/docs/element