Put London staging and prod pipelines on dashboard
We now have a full PaaS deployment in London. We already updated our wall monitor to include its `deploy-cloudfoundry` pipeline, but this link for people on support to see it on their computer was not updated.

I could not find a way to label each pipeline with which environment it is in. The wall monitor uses a Chrome plugin to do this.