Configure CircleCI to ignore gh-pages branch.
This adds a dummy CircleCI config file to the source directory so that
it gets included in the gh-pages branch. This will prevent CircleCI from
running a v1.0 build on the gh-pages branch, and erroring because it
can't detect any tests.

I've added a dummy docker and build step to prevent spurious errors in
the CircleCI UI. These don't actually cause a problem (the build is
still skipped), but are noise that could be confusing.