Fix the formatting of commands on the UAA users guide
This was squashing multi-line code blocks onto a single line, which
meant that when the commands were copy.pasted into the terminal, they no
longer worked (due to the `\`s in the command to handle the newlines
which were no longer present).

Adding a linebreak before these codeblocks prevents this happening.