Add reference to setting SLIM_DEV_DEPLOYMENT=true for upgrade stories
Since we have enabled the default of slim dev environments it is not a
true replica of production. As a result we are unable to test if any
downtime seen is due to the upgrade or the use of the slim dev
envionment. By enabling the full dev environment we should be more
confident in our testing of upgrades.