Add process for trial and abandoned accounts
We are now enforcing the 3 month period of trial accounts. this commit
adds a process that we can follow for repeatable results in collating
the data and contacting the org managers.

The notify scripts expect the data to be supplied as a JSON array, by
modifying the SQL queries we can achieve the required data format in the
export.