Update AWS access instructions to use assume-role
We now have the necessary config in place to support using assume-role
from the central gds-users account instead of needing individual users
in each account. This updates our docs to reflect this, and to link out
to the RE manual for how to set this up.

The actualy details of the available roles are kept in the
account-wide-terraform repo and linked out to from here to avoid
publishing the account IDs and roles in a publicly visible place.