Fix how to get RDS instances for a given org
The script we were using to list the RDS instances for a pen testing
was wrong and was not listing the guid of the service_isntance,
but the unique_id of the service offering (e.g. psql small).

The right way to do it would be:

 1. Query the guid of the service offering for the rds_broker.
 2. List the service_instances in the org with that guid in service_guid.

We update the script to support this.