Update Datadog to Prometheus
We are migrating away from Datadog to use Prometheus. This commit
updates the references and information that we have within the team
manual.

This needs to be done to ensure that the person on support has the right
information to hand in order to resolve issues in a timely manner.