Generalise ADRs that reference Datadog
We're removing Datadog, so these no longer apply. Instead of updating
them to refer to the equivelent metrics in the new system, I've reworded
them to generally describe the thing being measured so that they're
agnostic of the metrics system being used, and therefore shouldn't need
updating again if this changes.