Update google search to use richardtowers' account
Originally this was using Hector's account, but he's left now so the
integration has broken.

I've created a new integration and invited all the civil servants on the
team to be administrators.

I did this:

* Go to https://cse.google.co.uk/cse
* Click add
* Fill in details of https://team-manual.cloud.service.gov.uk/
* Copy the `cx` code out of the snippet it gives you and update it in
  `source/javascripts/application.js
* Add all the people on the team in the admins tab

Hopefully we shouldn't have to do this again because we've got several
people administering it now.