Document packages needed to run `make test` locally
As a new starter, I didn't find any documentation about what packages needed to
be installed for me to run `make test` locally, which lead to a long feedback
loop of waiting for travis to run the tests, making amendments and pushing
again. This commit introduces documentation for that case, based on the
packages installed by TravisCI in order to run the tests.