Add guidance for sending technical emails to tenants
In #165643581 we added a new tool to `paas-cf` for extracting the email
addresses of tenants from Cloud Foundry, so that we could send emails to an
accurate list of tenants. In this commit, we add guidance to the team manual on
how to use the tool and send technical emails.