Diagram of Certificate Authority rotation
I drew this diagram to analyse whether we really had to do quite so
many steps to rotate our CAs. I found a shorter path (the thick
arrows) but don't plan to implement it right now because of the
time-consuming testing that would be needed. However people have said
the diagram is helpful. It's only a first draft but this commit puts
it in to the page to help visual thinkers (like me!)