update table of zendesk tags
in https://www.pivotaltracker.com/n/projects/1275640/stories/175878658 we analyse our support tickets.

Our old taxonomy of tags was unecessarily complicated so the team have agreed to simplify the list and use a more mininal set of tags. The support engineer will use this set of tags to label the incoming tickets to facilitate analysis by the team.